Version 0.2.14
Implemented method vimside#ensime#swank#handle_no_response() which is call
by vimside#ensime#io#ping(timeout) when the response from the Ensime server
is empty which uses the previous checkin's Options
    scheduler-many-max-count-no-events (default 50)
    scheduler-events-max-count-no-events (default 50)
to ultimately restore the schedule to a "not expecting anything/steady" state
(a long time between pings). This simply assures that if there is some
bug in handling responses from the server, a fast ping rate will
ultimately stop and the system returned to steady-state.