Version 0.4.0
Big change was the addition of the ActWin (Action Window) which
is now the default window for both compiler errors (via Ensime or
SBT) and the uses-of-symbol-at-point command. Very configurable.
The quickfix window can still be use if desired via configuration.
Now Vimside supports either initialization with the Option files
'options_user.vim' and a project specific options file (and/) or
with Properties file 'vimside.properties' and a project specific
properties file. See the 'data/vimside/PROPERTIES' file for
available Properties (or, changing '.' to '-', Options).
Also, bumped (after a long time, the version number.